flask_version/requirements.txt
Flask
psycopg2-binary
python-dotenv
pytest
Flask-Testing
flask_sqlalchemy